feat: reference titles added
