+++ src/features/listing-builder/store/index.ts
@@ -72,6 +72,7 @@ const mergeListingWithInitialFormState = (
   references: (listing.references || []).map((e) => ({
     order: e.order,
     link: e.link,
+    title: e.title,
   })),
   publishedAt: listing.publishedAt,
   rewardAmount: listing.rewardAmount,
